feat: dynamic generation of RSS feeds via endpoint and getStaticPaths
